Change signature of CreateChangeset to return err last
This is a follow-up to #7328 and changes the signature of the
`CreateChangeset` method according to @tsenart's suggestions.

I'm going to leave the `CreateChangeset` method as it is right now,
instead of turning it into an `UpsertChangeset` method, since we don't
have BitbucketServer support yet and I'm not sure the interface would be
better.