all: fix some errcheck warnings
A new meditation for yuppie software engineers. Run errcheck on our codebase
and fix some of the warnings. 5 minutes a day will lead to a better, happier
and examined life.

Most fixes are just explicitly marking we are ignoring the error. This is the
case often since we are writing a response and there is no reasonable recourse
if it fails. Some changes handle the error.