e2e tests: Ensure repos are cloned before starting test suite
I have a hunch that not waiting for repos to be fully cloned is the
cause for the majority of our e2e flakiness (excluding
infrastructure-related flakiness).

I don't know whether this check was already there and got removed or
whether the architecture changed in the meantime (repos are added and
cloned async now), but the comment (`// Clone the repositories`) makes
it sound like this is what's supposed to happen in `ensureRepos`.

The test I use is to check for the revision selector to show up.