db: remove enabled/disabled from ReposListOptions
All repos are enabled, so these fields have been noops since 3.4. Had to
update all call sites for listing repos. If it specified `disabled: true` and
not `enabled: true` we know the result set is empty so the code is
updated. Otherwise the filter is a noop so we just remove specifying the
fields.