ci: golangci-lint all errors
The new-from-rev doesn't seem to be that trustworthy, since there are
instances of new code getting through. I have been sending out commits which
make us golangci-lint clean. We should be now be able to just run lint on
everything. Doing the cleanup found legitimate bugs, so this was a worthwhile
pursuit.