web: avoid flicker in site admin area for lazy-loaded pages
The pages in the site admin area are lazy-loaded components. Previously, the entire site admin area, including the sidebar, was hidden while the lazy-loaded component loaded. Now, only the content area (where the lazy-loaded component will be rendered) is hidden and a loading indicator shown. This significantly reduces the UI jitter in the site admin area.