a8n: Rename previewCampaignPlan to createCampaignPlan
This is a breaking change I'm proposing to make understanding the API a
lot easier.

I ran into the problem of having to explain what "previewing a campaign
plan" means multiple times in the past few days. First, when explaining
Automation-related concepts to Ryan and Kevin, and then when working on
`src-cli` to add the ability to create `CampaignPlan`s and `Campaign`s.

In the end I realized that `previewCampaignPlan` is simply not as clear
as `createCampaignPlan`, which is what the mutation actually does: it
creates a `CampaignPlan` that can then be shared, previewed and used to
create a `Campaign`.

As I wrote in the CHANGELOG.md: I think we should use the time we're
still in beta and behind a feature flag to make these changes.