gitserver: ListCloned filters out repos that do not belong on s… (#7969)
When rebalancing repositories across shards we can have a repository appearing
on a shard which doesn't belong there. ListCloned is used to report to Admins
which repos need to clone. This leads to not reporting any repos as needing
cloning when rebalancing. This commit updates ListCloned to filter out repos
that do not belong on a shard.

Note: Follow-up work is required to make our Purge job shard aware as well as
gracefully handle rebalancing.