web: include alwaysRender on blob error page (#8014)
Not including alwaysRender lead to some toolbar items not showing (such as the
download button). An example of a common error page is we don't render images.
However, we would still like to include the download button.

Note: We rely on the raw endpoint for the download functionality. For security
reasons this will always set the content-type to text/plain. This means when you
download the file it will be downloaded as a txt file. We can't support more
mime-types unless we can serve the assets on a different domain.