tracking-issue: Keep track of changes to planned work
This commit makes it so that issues that were originally planned in an
iteration (labelled with `planned/$MILESTONE`) don't disappear from the
tracking issue, and are instead crossed-off.

This allows us to keep track of changes to planned work throughout the
iteration.

A follow-up PR will update our documented process with an extra step of
labelling all issues in an approved tracking issue with the
`planned/$MILESTONE` label.