web: standardize top margin in site admin area (#8052)
Previously, most (but not all) site admin pages specified mt-3 or equivalent for a margin-top. Some pages did not, which meant that the were incorrectly flush with the global nav. Instead of having each site admin page specify mt-3, this changes it to just specify that in the parent component.