redispool: Wait SRC_REDIS_WAIT_FOR before quitting sysreq check (#8110)
* redispool: Wait SRC_REDIS_WAIT_FOR before quitting sysreq check

This commit makes our `redispool` sysreq check wait for `90s` by
default instead of `5s` and allows admins to set a `SRC_REDIS_WAIT_FOR`
environment variable to control this timeout if it turns out the new
default isn't enough.

Addresses part of #2904. The next step is make Redis start-up faster by
ensuring it maintains a small AOF file that is short to process on
startup (#3300).

Additionally, repo-updater also uses redis, but doesn't currently
integrate with the sysreq package. I'll be reviewing whether we still need
to use redis at all in it (it's currently used for caching GitHub API
responses). If we still need it, I'll run sysreq checks in repo-updater,
otherwise, I'll get rid of the redis dependency there.

* fixup! Update CHANGELOG

* fixup! Add check for redis cache and fix check func

Since we depend on Redis cache in many places as well, including
repo-updater which waits for the frontend to be ready before starting.

Additionally, the previous code didn't actually work because it reused
the same broken connection in each attempt, rather than trying to
establish a new one.