redis-pool: Use BGREWRITEAOF as sysreq command
This commit makes the sysreq check for redis store and cache run a
`BGREWRITEAOF` command rather than a simple `PING`. This ensures that
the AOF file doesn't grow unbounded due to frequent container restarts
(with the server image). The command itself is non-blocking and
best-effort.

The reason I didn't add this command to be ran from a Procline in the
server process manager is that it would require Redis to already by up,
and that would be substantially harder to wire together than this.

Fixes #3300