server: Redis persistence optimizations (#8114)
* redis-pool: Use BGREWRITEAOF as sysreq command

This commit makes the sysreq check for redis store and cache run a
`BGREWRITEAOF` command rather than a simple `PING`. This ensures that
the AOF file doesn't grow unbounded due to frequent container restarts
(with the server image). The command itself is non-blocking and
best-effort.

The reason I didn't add this command to be ran from a Procline in the
server process manager is that it would require Redis to already by up,
and that would be substantially harder to wire together than this.

Fixes #3300

* server: Redis persistence optimizations

This commit optimizes how we do Redis persistence in the server image by:

1. Upgrading Redis to version 5.0.7 so that we can use the
`aof-use-rdb-preamble` setting which substantially speeds up the server
image start-up time in certain scenarios (see #3300).
2. Making `redis-cache` truly ephemeral by not storing any snapshots or
append only log to disk.

* doc: Update Redis version

* Update CHANGELOG

* server: Always rewrite redis config

* fixup! ./dev/generate

* Update CHANGELOG.md

Co-Authored-By: Keegan Carruthers-Smith <keegan.csmith@gmail.com>

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>