docs: Remove /cache from ESLint example, mention cache in separate section (#8263)
Using the `/cache` directory when using Docker for Mac slows down the
actions by a lot. When removing the `/cache` the execution time for the
ESLint campaign when from 16min+ to 9min in a small repository
(codeintellify).

I suspect the reason for that is Docker for Mac's slow file syncing
between the container and the host.

So I decided to remove the mention of `cacheDirs` from the example
campaign and move it to its own section where it's still available for
"advanced users" that want to optimize their campaign run time.