a8n: Combine changeset and changeset events labels
We now read labels from the top level changeset event object as the
source of truth. We also look at any webhook events that have arrived
since the changeset was last updated in order to reflect recent updates.

The reason for this change is that timeline items are unreliable when
labels are added, removed or edited at the repository level.