a8n: Combine changeset and changeset events labels (#8281)
* a8n: Add migration to remove event changeset events

They will be regenerated on next sync. This prepares us to the move to
use the graphql node id as the key.

* a8n: Combine changeset and changeset events labels

We now read labels from the top level changeset event object as the
source of truth. We also look at any webhook events that have arrived
since the changeset was last updated in order to reflect recent updates.

The reason for this change is that timeline items are unreliable when
labels are added, removed or edited at the repository level.

* a8n: Leave out zero values in test

* a8n: Change interface of ChangesetEvents.Labels

* extsvc: Revert golden file

* extsvc: Revert test data