a8n: Send campaigns count in usage pings
This adds the total number of Automation campaigns in the database to
the update ping.

This is just one number by design, since for now we only want to capture
the number of campaigns per customer.

That's also why I simply defined one function in this package that
queries the database directly.

It doesn't use the the `a8n.Store` code, since that would require moving
the `Store` out of the `enterprise/internal/a8n` into the `internal/a8n`
package. So I went for this tradeoff (duplicate mention of `campaigns`
table here and in the store code).

What do you think?