a8n: Fix updating of a Campaign's name/description (#8437)
Before this change the code was broken: an attribute-only (name or
description) update to a Campaign was never persisted.

This previously worked but got lost in refactors while working on
"individual updates". The tests also didn't spot the problem, since they
only checked the returned struct.

The change here fixes the tests to load the campaign from the DB and
then fixes the now-failing tests by updating the campaign.