Always defined ExtWindows.activeWindow
Fixes https://github.com/sourcegraph/sourcegraph/issues/8519

This changes ExtWindows to initialize `activeWindow` in the constructor, and keep the same active window for its entire lifetime, rather than starting out with an undefined `activeWindow` and adding/removing it as editors are added/removed.

This makes it possible for extension to call methods on `sourcegraph.activeWindow`  (such as `showNotification()` and `showInputBox()`) on all pages where the extension host is active, rather than solely on code view pages.