campaigns: Changeset syncing should only happen in repo updater
We now only have one instance of campaigns.ChangesetSyncer which is
instantiated during startup of enterprise/repo-updater.

Instead of creating a syncer from the frontend we now make an HTTP RPC
call to repo-updater to enqueue the sync. The actual enqueing process has
not been implemented yet and will be done in another PR, this change just
refactors the code to make it possible.