campaigns: Introduce HTTP endpoints for changeset syncing (#8590)
* campaigns: Changeset syncing should only happen in repo updater

We now only have one instance of campaigns.ChangesetSyncer which is
instantiated during startup of enterprise/repo-updater.

Instead of creating a syncer from the frontend we now make an HTTP RPC
call to repo-updater to enqueue the sync. The actual enqueing process has
not been implemented yet and will be done in another PR, this change just
refactors the code to make it possible.

* campaigns: Force blocking sync in certain cases.

When creating or closing changesets we perform a blocking sync.

We're making the tradeoff that the user my experience a rate
limit issue but the advantage is much simpler code and the invariant
that we'll always have performed an initial sync before persisting a
changeset to the db.

If we had performed the sync without blockng we'd temporarily have an
unsynced changeset. We'd have to check for this everywhere that we read
changesets. Also, the remote changeset may not exist and the error would
be difficult to display to the user.

We also decided not to make an RPC requests to repo-updater as we need
to keep the changes within one DB transaction.

* campaigns: Remove old code

* campaigns: More old code

* repoupdater: Update comment

* campaigns: Revert change

* campaigns: Typo