repo-updater: Initial work on more intelligent changeset syncer
Currently does a lot of logic in code which will be moved to the DB.

The core idea is that we look at the difference between when we last
synced and when the changeset was last changed remotely. We use that
number to decide when next to sync. We back off linearly up to a maximum
time.

For example, if we last synced at 10am and the last change was at 9:55am
(5 minutes) our next sync will be 10:05. So a 5 minute gap will lead us
to schedule the next update in 10 minutes. A 20 minute gap ould lead us
to schedule the next update in 40 minutes.

The overall schedule is recomputed every 2 minutes currently.