Fix unmarshaling and displaying of gitserver CreateCommitFromPatchError
Before this change I ran into this error message:

    frontend | WARN decoding gitserver create-commit-from-patch response, err: json: cannot unmarshal object into Go struct field CreateCommitFromPatchError.Error.Err of type error

This commit changes the field from an `error` to a `string`, which can
be marshaled/unmarshaled by the json package.

I don't see where we need to satisfy the `Unwrap()` interface and
preserve the original error, so in order to make this work, I got rid of
it.