search: increase default indexed-search timeout to 10s (#8720)
Previously zoekt would timeout after 3s by default. This is a very
conservative time, but on an established instance is enough. However, we have
had two new customers this week run into unstable result counts due to this
timeout. They needed to tune there resource allocations to fix the underlying
issue, but the user experience was poor and confusing. So we increase the
timeout to make it clear that search is slow rather than broken.

This new timeout is far greater than we expect zoekt to take. The reason we
use this timeout is we would rather be slow than inaccurate. Additionally if a
user opts in to some timeout settings, those override the default.