campaigns: Introduce heuristic syncing of changesets (#8653)
* repo-updater: Initial work on more intelligent changeset syncer

Currently does a lot of logic in code which will be moved to the DB.

The core idea is that we look at the difference between when we last
synced and when the changeset was last changed remotely. We use that
number to decide when next to sync. We back off linearly up to a maximum
time.

For example, if we last synced at 10am and the last change was at 9:55am
(5 minutes) our next sync will be 10:05. So a 5 minute gap will lead us
to schedule the next update in 10 minutes. A 20 minute gap ould lead us
to schedule the next update in 40 minutes.

The overall schedule is recomputed every 2 minutes currently.

* migrations: Add external_upated_at to changesets

* campaigns: Update external_updated_at during changeset sync

* campaigns: Add ListChangesetSyncHeuristics to campaigns store

For now it returns sync timing data which we use to decide when to next
sync a changeset.

* campaigns: Use ListChangesetSyncHeuristics in syncer

* campaigns: Remove dead code

* campaigns: Changeset syncer can now be sent high priority items

* campaigns: Add a test for nextSync

* campaigns: Updated time should be max of updated_at and latest_event

* campaigns: Fix lastChange calc

The latest change should be the most recent of either the last event
we've seen on their update.

* campaigns: Remove logging, add comments

* campaigns: Fix comment

* dev: Rerun generate

* campaigns: Simplify changeset backoff

* Update enterprise/internal/campaigns/store.go

Co-Authored-By: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>

* Update cmd/repo-updater/repos/bitbucketserver.go

Co-Authored-By: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>

* campaigns: Initial changeset schedule performs limited retries

* campaigns: No need to reuse variable

* campaigns: Extra test case

* campaigns: More comments and a rename

* repo-updater: Extract unixMilliToTime

* campaigns: Remove db retry logic

We don't do it anywhere else in the code and we'll try again later in
the main loop

* Update enterprise/internal/campaigns/store.go

Co-Authored-By: Thorsten Ball <mrnugget@gmail.com>

* Update enterprise/internal/campaigns/store.go

Co-Authored-By: Thorsten Ball <mrnugget@gmail.com>

* campaigns: Move ChangesetSyncHeuristics from store.go to types.go

* campaigns: Remove clock as it is not used

* campaigns: Refactor nextSync to take ChangesetSyncHeuristics

* campaigns: Fix listChangesetSyncHeuristicsQuery

We need to user a LEFT JOIN as we could have changesets with no events

* Apply suggestions from code review

Co-Authored-By: Thorsten Ball <mrnugget@gmail.com>
Co-Authored-By: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>

* repo-updater: Renaming

* campaigns: Test that ExternalUpdatedAt is populated

Co-authored-by: ᴜɴᴋɴᴡᴏɴ <u@gogs.io>
Co-authored-by: Thorsten Ball <mrnugget@gmail.com>