Add a not-null constraint to campaigns.description
This is just a clean up, because @eseliger and I found out yesterday
that we don't have a `NOT NULL` on this column and, at the same time,
have no `NullString` scanning in place.

Technically it's not an issue, since we always write blank strings, but
this just makes it more explicit.