CHANGELOG entries for 3.12.{8,9,10} and 3.13.2 (#8968)
We are about to release 3.12.10 and 3.13.2 which both contain the same
fixes. I have included this note, since in the past we haven't often
backported to older releases.

I am leaving the CHANGELOG entries for 3.14 in since users upgrading from
older 3.13 patch releases would be interested in reading them. I don't know
what our policy is around this, but this seems like a reasonable approach.

I also noticed 3.12.8 and 3.12.9 were missing entries. I added those in.