campaigns: Changeset resolver now uses derived fields  (#8986)
* campaigns: Bitbucket fields should be pointers

This caused an interesting bug. When we load bitbucket changesets from
our db we store it in the interface{} Metadata field as a
*bitbucketserver.Activity. When we loaded it from the codehost we stored
it as a value, not a pointer. In other parts of the code we used a type
switch to grab the value back out that assumed it was always a pointer.
One of these type switches was used to get the timestamp of the event
and would fall back to the zero value if the switch didn't match. This
resulted in us always having a zero timestamp, but onlu when loaded from
the code host.

* campaigns: Changeset resolver now used derived fields

Instead of computing values at query time