gitserver: support cloning repos with a customFetchCmd
The existing customer which uses customFetchCmd manually cloned the repository
into gitserver, so the support for customFetchCmd previously didn't take into
account the need to clone.

We only know how to fetch these repos, so instead of calling "git clone" we
instead init an empty repository and then run fetch. I am unsure the wider
implications of missing a remote URL in the git config / etc. However, the
testing I did should unblock the customer relying on the need to clone.