campaigns: Only load GitHub checks in top level commit
When fetching a PR from GitHub we fetch commits in two sections of the
GraphQL request:

1. The most recent commit on the PR
2. All commits as part of timeline items

Previously, we requested both types of commits using the same GraphQL
fragment which included all check data (checkruns, suites and commit
status). We only need this for the most recent commit as that is the one
we use to decide on the current build status.

If we only request check data for the most recent commit it has a large
impact on our API rate limiting as it drops our request cost from 28 down
to 3.

Backwards compatibility:

We store the raw graphql response in two places in the DB. In metdata on
a ChangeSet and metadata in a ChangesetEvent.

The data stored in the Changeset has not changed, it has just been
renamed in code.

The data in the event has had its check data removed. Existing data,
when unmarshalled, will simply have no fields to marshal into which is
fine and the old fields were never read so they are safe to remove.