conf: remove unused (and racy) migration
This migration existed to merge the critical configuration into the site
configuration as we remoed the Management Console in v3.11.

The migration is no longer needed as anyone upgrading will have already
ran this OR will need to visit 3.11 first per our upgrade guidelines,
and additionally this migration is racey when specifying site or critical
configuration via a file as both get applied at te same time which leads
to ones' edits being rejected. In practice, this is not an issue as the
migration is eventually successful in any case -- it just means the server
may restart needlessly due to "the migration failing".

Fixes #9019