repo-updater: introduce excluder (#9047)
This is a first and minimal attempt at factoring out the common logic in our
exclude clause. It is just targetting names and IDs, since that is the only
thing common across all code hosts.

In future this can be evolved into taking a common schema object, etc. This
commit only applies it to GitHub, rest of the sources will also use it in
follow-up commits.