Remove execution-related fields from CampaignJobs
This removes all the fields on CampaignJob that are not used anymore:

* FinishedAt
* StartedAt
* Error
* Description

With #8008 those became useless.

What this does is remove the fields but leave the GraphQL API intact.
That means we can freely merge it without breaking anything.

The changes in #9068 will need to be rebased on this here.