zoekt: prometheus metrics and debug logs
This includes changes to zoekt which adds request logs and prometheus
metrics. This is to aid debugging of performance issues.

We add some points to the "slow search" troubleshooting scenario.

Zoekt now respects SRC_LOG_LEVEL and won't output "listening on" unless debug
logs are enabled. As such we can remove the "grep -v" we have in the server
procfile entry.