search: succinct TextPatternInfo for searchFilesInRepos traces (#9151)
We just put the query string in the title instead of the whole TextPatternInfo
into the title. We then log the TextPatternInfo. We also log the parsed query
which should help with a current problem we are facing.

Additionally searcher.protocol.PatternInfo is a type that is duplicated with
search.TextPatternInfo. We need to unify them in a later commit, but for now
it is easier to copy the succinct string function for PatternInfo to
TextPatternInfo. This will produce much easier to read log output, since
usually most fields are just the default empty value.

This change was motivated by struggling to read traces when investigating a
search issue.