trace: remove ancestors from family name (#9195)
This "flattens" the net/trace debug page. I find this much more readable and
for common "leaf" traces this makes it much easier to understand its
performance. For example we can now get a good idea on how "sql" is
performing overall.

To compensate for the missing ancestor state, it is logged as an event in the
trace.