frontend: use endpoint port if prometheus.io/port annotation is missing (#9216)
Previously we excluded indexed-search from scraping. Instead we
use the same behaviour as prometheus and use the endpoint port of the
annotation is missing.

Note: we still won't proxy to zoekt-indexserver. This would require us to
watch the pods as well, since we don't have any services setup for
zoekt-indexserver.