dev: handle failures during docsite binary download (#9234)
If docsite.sh is killed while fetching a binary a partial binary will exist at
target. Following runs will try and fail to execute the binary. We now
download to a temporary location and move it into place, so if we fail we will
retry the download. Additionally we just always ensure the binary is
executable.