repo-updater: use value instead of pointer for configuredRepo2
This prevents any race conditions around us updating the pointers we had to
configuredRepo2. Additionally it may reduce garbage we produce in the heap.

Previosly we added a coarse lock and adjusted the mutex to an RWMutex. The
coarse lock went from holding during an O(n) operation to an O(nlogn).
Additionally RWMutex was inappropriate since RLock would only happen very
rarely and there would be lots of quick Lock calls. RWMutex is not appropriate
for that pattern.