Fix: double quotes not being handled properly in interactive mode
Now, we will always wrap the content and message filters in double quotes,
and escape any double quotes within the input. For all other filters, when
double quotes are found in the input, we will escape them and wrap the entire input in
double quotes. If spaces are found in the input, we will also wrap the entire input in
double quotes.