Hard exit when user hits Ctrl-C a second time (#170)
Right now we don't propagate the context to everything (most notably:
API requests). That means the program gets stuck when it's running a
portion of the code that doesn't use the context and the user hits
Ctrl-C.

To avoid that we listen for a second Ctrl-C and then exit with the
interrupt code 2.