Improve query speed for extension registry
When looking into another slow request, I found that this query
made up 32% of the total execution time of queries,
according to pghero. To my understanding of SQL, this should be equal
with postgresql. I tested it using explain analyze and it's
about 5x as fast (rudimentary average calculation over 5 manual runs).