Handle dismissed GitHub review events correctly
What GitHub does when a review is dismissed is to:

* update the original review and set its state to "dismissed"
* emit an "review dismissed" event

Remove all of the reviews of the author whose review was dismissed from
the overall review state.

Previously we not only deleted the last review by the same author if we
found a "review with state 'dismissed'" event but we also deleted the
reviews when we found a "review dismissed" event, which can lead to
double-deleted reviews.