Apply bandaid to symbols restarting in infinite loop
See my previous cries for help for context:

- https://sourcegraph.slack.com/archives/CHPC7UX16/p1585734381060100
- https://sourcegraph.slack.com/archives/C07KZF47K/p1585635858081000
- https://sourcegraph.slack.com/archives/C07KZF47K/p1583916442139400

Short version: `symbols` keeps restarting for me ever since I made that
change in #8912.

My current theory is that the existing check (which I'm changing here)
doesn't work for symbols because:

1. In `./dev/go-install` we use `go install` to build the symbols binary
2. Then we restart `symbols`
3. `symbols` "starts" by doing its own compilation, which changes a file
   in `cmd/symbols/*`
4. That causes `handle-change.sh` to restart the service that's
   currently compiling???

Yes, it's not complete...

Anyway: this here fixes it for me locally and since nobody else

1. complained about the same bug
2. asked for my changes in #8912

I'm hereby reverting the changes in #8912 for symbols.